+++ cms/src/features/posts/Posts.tsx
@@ -1,7 +1,10 @@
+import PostForm from "./components/PostForm";
+
 const Posts = () => {
   return (
     <>
       <h2>Posts</h2>
+      <PostForm />
     </>
   );
 };